requirements.txt
aiogram==3.1.1
python-dotenv
openai
google-generativeai
